requirements.txt
pyobjc-framework-Vision>=10.0
pyobjc-framework-Quartz>=10.0
